backend/requirements.txt
# Core framework
fastapi
uvicorn[standard]

# Async HTTP and event loop support
httpx
nest_asyncio
requests-cache

# Data validation and configuration
pydantic
pydantic[email]  # for email-specific validation

# Environment and logging
python-dotenv
loguru

# Databases and storage
sqlalchemy
pymongo
llama-index # for reordering

# Multipart/form-data handling
python-multipart

# Security
passlib[bcrypt]
python-jose[cryptography]

# PDF and document handling
pypdf
PyMuPDF
pdf2markdown4llm

# External APIs and clients
openai~=1.93.0
weaviate-client==4.16.6
neo4j
protobuf~=5.29.0 # warning fix
requests
selenium

# Data processing
pandas
numpy>=1.21.0
yake

# Evaluation and NLP
evaluate
datasets
nltk
absl-py
rouge-score
bert_score
rapidfuzz>=3.6.2 # mainly for levenshtein
ragas

# Model inference (e.g., BAAI/llm-embedder)
torch>=2.6
transformers>=4.46.3
FlagEmbedding

# Optional: Safetensors for secure model loading
safetensors  # Use this for secure model loading if needed
